1. fixed overflow issue on time slider
2. fixed time slider jolting in LIVE mode

<==== Mercurial data ====>
Branch: master

+++ client/src/ui/videoitem/timeslider.h
@@ -14,7 +14,7 @@ class TimeSlider : public GraphicsWidget
     Q_OBJECT
     Q_PROPERTY(qint64 currentValue READ currentValue WRITE setCurrentValue NOTIFY currentValueChanged)
     Q_PROPERTY(qint64 maximumValue READ maximumValue WRITE setMaximumValue NOTIFY maximumValueChanged)
-    Q_PROPERTY(float scalingFactor READ scalingFactor WRITE setScalingFactor NOTIFY scalingFactorChanged)
+    Q_PROPERTY(double scalingFactor READ scalingFactor WRITE setScalingFactor NOTIFY scalingFactorChanged)
     Q_PROPERTY(qint64 viewPortPos READ viewPortPos WRITE setViewPortPos)
 
 public:
@@ -27,7 +27,7 @@ public:
     qint64 minimumValue() const;
     qint64 maximumValue() const;
     qint64 currentValue() const;
-    float scalingFactor() const;
+    double scalingFactor() const;
 
     bool isMoving() const;
     void setMoving(bool b);
@@ -44,7 +44,7 @@ public Q_SLOTS:
     void setMinimumValue(qint64 value);
     void setMaximumValue(qint64 value);
     void setCurrentValue(qint64 value);
-    void setScalingFactor(float factor);
+    void setScalingFactor(double factor);
 
     inline void zoomIn()
     { setScalingFactor(scalingFactor() + 1); }
@@ -70,7 +70,7 @@ private Q_SLOTS:
     void onWheelAnimationFinished();
 
 private:
-    float delta() const;
+    double delta() const;
 
     qint64 fromSlider(int value);
     int toSlider(qint64 value);
@@ -89,14 +89,13 @@ private:
     qint64 m_currentValue;
     qint64 m_viewPortPos;
 
-    float m_scalingFactor;
+    double m_scalingFactor;
     bool m_isUserInput;
     int m_delta;
 
     QPropertyAnimation *m_animation;
     bool m_centralise;
     qint64 m_minimumRange;
-
 
     friend class TimeLine;
 };
